+++ src/app/framework/native/components/product/list/index.ts
@@ -1,1 +1,1 @@
-export * from './list-products';
+export * from './product-list';
