+++ src/app/framework/native/components/product/list/product-item/index.ts
@@ -1,0 +1,1 @@
+export * from './product-item';
