+++ src/app/framework/native/containers/private/product/list/index.ts
@@ -1,1 +1,1 @@
-export * from './list-products.view';
+export * from './product-list.view';
