+++ src/app/framework/native/containers/private/product/list/product-list.hook.ts
@@ -1,5 +1,5 @@
 import {} from 'react-native';
 
-export function useListProductsModel() {
+export function useProductListModel() {
   return {};
 }
