+++ src/app/framework/native/containers/private/product/list/product-list.style.ts
