+++ src/app/framework/native/containers/private/product/list/product-list.type.ts
@@ -1,7 +1,7 @@
 import type {NativeStackScreenProps} from '@react-navigation/native-stack';
 import {BottomTabsParamList} from '@native/navigation';
 
-export type ListProductsProps = NativeStackScreenProps<
+export type ProductListProps = NativeStackScreenProps<
   BottomTabsParamList,
-  'ListProducts'
+  'ProductList'
 >;
